refactor(API Views): create group discard mixin to be able to share it between routers
